refactor: Shared 컴포넌트를 SharedPage 컴포넌트로 변경
