--- src/components/Shared/index.js
@@ -1,1 +1,0 @@
-export { default as Shared } from './Shared';
